+++ lib/types.ts
@@ -151,9 +151,8 @@ export interface InnerFormProps {
 
 export type FileInputResponse = {
   name: string;
-  file: File;
-  src: FileReader;
-  [key: string]: string | File | FileReader;
+  file: string;
+  [key: string]: string | number | File | FileReader;
 };
 
 export interface AuthenticatedUser {
@@ -187,6 +186,20 @@ export interface CrudTemplateResponse {
   };
 }
 
+// Submission
+export interface SubmissionRequestBody {
+  [key: string]: string | string[] | FileInputResponse | FileInputResponse[];
+}
+
+export type SubmissionParsedRequest = {
+  fields: Record<string, Response>;
+  files: Record<string, ProcessedFile | ProcessedFile[]>;
+};
+
+export interface ProcessedFile {
+  name: string;
+  buffer: Buffer;
+}
 // CRUD Operations for Organizations
 export type Organization = {
   organizationID: string;
@@ -244,6 +257,8 @@ export type MiddlewareRequest = (
 export interface MiddlewareProps {
   formID?: string;
   session?: Session;
+  email?: string;
+  temporaryToken?: string;
 }
 
 // User Types
@@ -255,3 +270,8 @@ export interface AuthenticatedUser extends User {
 export interface ExtendedSession extends Session {
   user?: AuthenticatedUser;
 }
+export type FormOwner = {
+  id: number;
+  email: string;
+  active: boolean;
+};